Removed Msg->reply_to because it held data for an older version of the code, we no longer do it this way

+++ webcit/tcp_sockets.c
@@ -275,8 +275,11 @@ void FlushReadBuf (void)
 		pche = pch + len;
 		if (WCC->ReadPos != pche)
 		{
-			syslog(1, "ERROR: somebody didn't eat his soup! Remaing Chars: %ld [%s]\n", 
-				pche - WCC->ReadPos, pche);
+			syslog(1,
+				"ERROR: somebody didn't eat his soup! Remaing Chars: %ld [%s]\n", 
+				(long)(pche - WCC->ReadPos),
+				pche
+			);
 			syslog(1, 
 				"--------------------------------------------------------------------------------\n"
 				"Whole buf: [%s]\n"
